restructuring a little bit for clarity
