+++ com.brgs.orm/Azure/Helpers/AzureStorageHelpers.cs
