+++ com.brgs.orm/Azure/Helpers/TableEntityBuilder.cs
