+++ com.brgs.orm/Azure/Helpers/TableEntityDecoder.cs
